+++ packages/jupyter-ai-magics/pyproject.toml
@@ -42,6 +42,7 @@ all = [
     "huggingface_hub",
     "ipywidgets",
     "langchain_anthropic",
+    "langchain_aws",
     "langchain_cohere",
     "langchain_google_genai",
     "langchain_mistralai",
@@ -64,9 +65,9 @@ ollama = "jupyter_ai_magics:OllamaProvider"
 openai = "jupyter_ai_magics.partner_providers.openai:OpenAIProvider"
 openai-chat = "jupyter_ai_magics.partner_providers.openai:ChatOpenAIProvider"
 azure-chat-openai = "jupyter_ai_magics.partner_providers.openai:AzureChatOpenAIProvider"
-sagemaker-endpoint = "jupyter_ai_magics:SmEndpointProvider"
-amazon-bedrock = "jupyter_ai_magics:BedrockProvider"
-amazon-bedrock-chat = "jupyter_ai_magics:BedrockChatProvider"
+sagemaker-endpoint = "jupyter_ai_magics.partner_providers.aws:SmEndpointProvider"
+amazon-bedrock = "jupyter_ai_magics.partner_providers.aws:BedrockProvider"
+amazon-bedrock-chat = "jupyter_ai_magics.partner_providers.aws:BedrockChatProvider"
 qianfan = "jupyter_ai_magics:QianfanProvider"
 nvidia-chat = "jupyter_ai_magics.partner_providers.nvidia:ChatNVIDIAProvider"
 together-ai = "jupyter_ai_magics:TogetherAIProvider"
@@ -74,7 +75,7 @@ gemini = "jupyter_ai_magics.partner_providers.gemini:GeminiProvider"
 mistralai = "jupyter_ai_magics.partner_providers.mistralai:MistralAIProvider"
 
 [project.entry-points."jupyter_ai.embeddings_model_providers"]
-bedrock = "jupyter_ai_magics:BedrockEmbeddingsProvider"
+bedrock = "jupyter_ai_magics.partner_providers.aws:BedrockEmbeddingsProvider"
 cohere = "jupyter_ai_magics.partner_providers.cohere:CohereEmbeddingsProvider"
 mistralai = "jupyter_ai_magics.partner_providers.mistralai:MistralAIEmbeddingsProvider"
 gpt4all = "jupyter_ai_magics:GPT4AllEmbeddingsProvider"
